+++ src/kirby/components/grid/grid-helper.service.tns.ts
@@ -1,0 +1,26 @@
+import { Injectable } from '@angular/core';
+import { screen } from 'platform';
+import { OrientationChangedEventData } from 'application';
+import * as app from 'application';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class GridHelperService {
+  currentScreenWidth: number;
+
+  constructor() { }
+
+  onInit(callback: () => void) {
+    this.currentScreenWidth = screen.mainScreen.widthDIPs;
+    app.on(app.orientationChangedEvent, (args: OrientationChangedEventData) => {
+      if (this.currentScreenWidth === screen.mainScreen.widthDIPs) {
+        this.currentScreenWidth = screen.mainScreen.heightDIPs;
+      } else {
+        this.currentScreenWidth = screen.mainScreen.widthDIPs;
+      }
+      callback();
+    });
+  }
+
+}
