+++ src/kirby/components/grid/grid-helper.service.ts
@@ -1,0 +1,13 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class GridHelperService {
+  currentScreenWidth: 800;
+
+  constructor() { }
+
+  onInit(callback: () => void) {
+  }
+}
