+++ src/kirby/scss/scss-helper.tns.ts
@@ -1,3 +1,5 @@
+import { Color } from 'tns-core-modules/color';
+
 /**
  * Always remember to keep this class and _native-variables.scss in sync when needed
  */
@@ -13,9 +15,9 @@ export class ScssHelper {
   public static BREAKPOINT_CARD_M = 460;
   public static BREAKPOINT_CARD_L = 820;
 
-  // iOS specific values, not used on web
-  public static SHADOW_OPACITY = null;
-  public static SHADOW_RADIUS = null;
-  public static SHADOW_OFFSET_Y = null;
-  public static SHADOW_COLOR = null;
+  // iOS specific values
+  public static SHADOW_OPACITY = 0.5;
+  public static SHADOW_RADIUS = 6.0;
+  public static SHADOW_OFFSET_Y = 2.0;
+  public static SHADOW_COLOR = new Color('#1c1c1c');
 }
